--- src/main/java/com/triploguebe/controller/FriendController.java
@@ -1,4 +1,0 @@
-package com.triploguebe.controller;
-
-public class FriendController {
-}
