--- src/main/java/com/triploguebe/dto/FriendRequestDto.java
@@ -1,4 +1,0 @@
-package com.triploguebe.dto;
-
-public class FriendRequestDto {
-}
