--- src/main/java/com/triploguebe/entity/Friendship.java
@@ -1,4 +1,0 @@
-package com.triploguebe.entity;
-
-public class Friendship {
-}
