+++ src/main/java/com/triploguebe/friend/controller/FriendController.java
@@ -1,0 +1,4 @@
+package com.triploguebe.friend.controller;
+
+public class FriendController {
+}
