+++ src/main/java/com/triploguebe/friend/dto/FriendRequestDto.java
@@ -1,0 +1,4 @@
+package com.triploguebe.friend.dto;
+
+public class FriendRequestDto {
+}
