+++ src/main/java/com/triploguebe/friend/dto/FriendResponseDto.java
@@ -1,4 +1,4 @@
-package com.triploguebe.dto;
+package com.triploguebe.friend.dto;
 
 public class FriendResponseDto {
 }
