+++ src/main/java/com/triploguebe/friend/entity/Friendship.java
@@ -1,0 +1,4 @@
+package com.triploguebe.friend.entity;
+
+public class Friendship {
+}
