+++ src/main/java/com/triploguebe/friend/repository/FriendshipRepository.java
@@ -1,0 +1,4 @@
+package com.triploguebe.friend.repository;
+
+public class FriendshipRepository {
+}
