+++ src/main/java/com/triploguebe/friend/service/FriendService.java
@@ -1,0 +1,4 @@
+package com.triploguebe.friend.service;
+
+public class FriendService {
+}
