--- src/main/java/com/triploguebe/repository/FriendshipRepository.java
@@ -1,4 +1,0 @@
-package com.triploguebe.repository;
-
-public class FriendshipRepository {
-}
