--- src/main/java/com/triploguebe/service/FriendService.java
@@ -1,4 +1,0 @@
-package com.triploguebe.service;
-
-public class FriendService {
-}
